requirements.txt
tf-keras
tensorflow
scikit-learn
transformers
datasets
google-generativeai